Extracting Go language frontend into own module
