+++ build.gradle.kts
@@ -27,10 +27,11 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
     // built-in
-    //java
+    java
     //`java-library`
     jacoco
     signing
+    `maven-publish`
 
     id("org.sonarqube") version "3.3"
     id("com.diffplug.spotless") version "6.2.0"
@@ -52,6 +53,8 @@ subprojects {
     apply(plugin = "jacoco")
     apply(plugin = "org.jetbrains.kotlin.jvm")
     apply(plugin = "com.diffplug.spotless")
+    apply(plugin = "maven-publish")
+    apply(plugin = "signing")
 
     repositories {
         mavenCentral()
@@ -67,9 +70,75 @@ subprojects {
         }
     }
 
+    publishing {
+        publications {
+            create<MavenPublication>(name) {
+                from(components["java"])
+
+                pom {
+                    url.set("https://github.com/Fraunhofer-AISEC/cpg")
+                    licenses {
+                        license {
+                            name.set("The Apache License, Version 2.0")
+                            url.set("http://www.apache.org/licenses/LICENSE-2.0.txt")
+                        }
+                    }
+                    developers {
+                        developer {
+                            id.set("oxisto")
+                            organization.set("Fraunhofer AISEC")
+                            organizationUrl.set("https://www.aisec.fraunhofer.de")
+                        }
+                    }
+                    scm {
+                        connection.set("scm:git:git://github.com:Fraunhofer-AISEC/cpg.git")
+                        developerConnection.set("scm:git:ssh://github.com:Fraunhofer-AISEC/cpg.git")
+                        url.set("https://github.com/Fraunhofer-AISEC/cpg")
+                    }
+                }
+            }
+        }
+
+        repositories {
+            maven {
+                url = uri("https://oss.sonatype.org/service/local/staging/deploy/maven2")
+
+                credentials {
+                    val mavenCentralUsername: String? by project
+                    val mavenCentralPassword: String? by project
+
+                    username = mavenCentralUsername
+                    password = mavenCentralPassword
+                }
+            }
+        }
+    }
+
+    signing {
+        val signingKey: String? by project
+        val signingPassword: String? by project
+
+        useInMemoryPgpKeys(signingKey, signingPassword)
+
+        setRequired({
+            gradle.taskGraph.hasTask("publish")
+        })
+
+        sign(publishing.publications[name])
+    }
+
+    tasks.withType<GenerateModuleMetadata> {
+        enabled = false
+    }
+
     tasks.withType<JavaCompile> {
         sourceCompatibility = "11"
         targetCompatibility = "11"
+    }
+
+    java {
+        withJavadocJar()
+        withSourcesJar()
     }
 
     tasks.withType<JacocoReport> {
@@ -120,7 +189,7 @@ subprojects {
  */
 """
 
-    var headerWithHashes = """#
+    val headerWithHashes = """#
 # Copyright (c) ${"$"}YEAR, Fraunhofer AISEC. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
