+++ cpg-core/src/main/java/de/fraunhofer/aisec/cpg/graph/TypeManager.java
@@ -27,7 +27,6 @@ package de.fraunhofer.aisec.cpg.graph;
 
 import de.fraunhofer.aisec.cpg.frontends.LanguageFrontend;
 import de.fraunhofer.aisec.cpg.frontends.cpp.CXXLanguageFrontend;
-import de.fraunhofer.aisec.cpg.frontends.golang.GoLanguageFrontend;
 import de.fraunhofer.aisec.cpg.frontends.java.JavaLanguageFrontend;
 import de.fraunhofer.aisec.cpg.frontends.typescript.TypeScriptLanguageFrontend;
 import de.fraunhofer.aisec.cpg.graph.declarations.RecordDeclaration;
@@ -56,6 +55,7 @@ public class TypeManager {
 
   private static Class<?> llvmClass = null;
   private static Class<?> pythonClass = null;
+  private static Class<?> goClass = null;
 
   static {
     try {
@@ -66,6 +66,10 @@ public class TypeManager {
     try {
       pythonClass =
           Class.forName("de.fraunhofer.aisec.cpg.frontends.python.PythonLanguageFrontend");
+    } catch (ClassNotFoundException ignored) {
+    }
+    try {
+      goClass = Class.forName("de.fraunhofer.aisec.cpg.frontends.golang.GoLanguageFrontend");
     } catch (ClassNotFoundException ignored) {
     }
   }
@@ -535,7 +539,9 @@ public class TypeManager {
       return Language.JAVA;
     } else if (frontend instanceof CXXLanguageFrontend) {
       return Language.CXX;
-    } else if (frontend instanceof GoLanguageFrontend) {
+    } else if (frontend != null
+        && goClass != null
+        && goClass.isAssignableFrom(frontend.getClass())) {
       return Language.GO;
     } else if (frontend != null
         && pythonClass != null
