+++ cpg-core/src/main/java/de/fraunhofer/aisec/cpg/passes/CallResolver.java
@@ -1251,7 +1251,8 @@ public class CallResolver extends Pass {
 
   /**
    * Adds the resolved default template arguments recursively to the templateParameter list of the
-   * ConstructExpression until a fixpoint is reached e.g. template&lt;class Type1, class Type2 = Type1&gt;
+   * ConstructExpression until a fixpoint is reached e.g. template&lt;class Type1, class Type2 =
+   * Type1&gt;
    *
    * @param constructExpression
    * @param template
