+++ cpg-core/src/testFixtures/java/de/fraunhofer/aisec/cpg/BaseTest.java
