+++ cpg-language-go/src/main/golang/build.sh
