+++ cpg-language-go/src/main/golang/declarations.go
