+++ cpg-language-go/src/main/golang/language.go
