+++ cpg-language-go/src/main/golang/lib/cpg/main.go
