+++ cpg-language-go/src/main/golang/location.go
