+++ cpg-language-go/src/main/golang/node.go
