+++ cpg-language-go/src/main/golang/scope.go
