+++ cpg-language-go/src/test/resources/golang-modules/awesome.go
