+++ cpg-language-go/src/test/resources/golang-modules/cmd/awesome/main.go
