+++ cpg-language-go/src/test/resources/golang-modules/go.mod
