+++ cpg-language-go/src/test/resources/golang-modules/util/stuff.go
