+++ cpg-language-go/src/test/resources/golang/call.go
