+++ cpg-language-go/src/test/resources/golang/comment.go
