+++ cpg-language-go/src/test/resources/golang/construct.go
