+++ cpg-language-go/src/test/resources/golang/field.go
