+++ cpg-language-go/src/test/resources/golang/for.go
