+++ cpg-language-go/src/test/resources/golang/function.go
