+++ cpg-language-go/src/test/resources/golang/go.mod
