+++ cpg-language-go/src/test/resources/golang/if.go
