+++ cpg-language-go/src/test/resources/golang/literal.go
