+++ cpg-language-go/src/test/resources/golang/struct.go
