+++ cpg-language-go/src/test/resources/golang/switch.go
