+++ cpg-neo4j/build.gradle.kts
@@ -58,6 +58,7 @@ dependencies {
     api(project(":cpg-core"))
     api(project(":cpg-language-llvm"))
     api(project(":cpg-language-python"))
+    api(project(":cpg-language-go"))
 
     implementation("org.apache.logging.log4j:log4j-slf4j18-impl:2.17.0")
