+++ settings.gradle.kts
@@ -8,6 +8,7 @@ include(":cpg-analysis")
 include(":cpg-neo4j")
 include(":cpg-language-llvm")
 include(":cpg-language-python")
+include(":cpg-language-go")
 include(":cpg-console")
 
 gradleEnterprise {
